backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
google-genai==1.14.0
pydantic==2.10.6
pydantic-settings==2.7.1
websockets==14.2
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
